[FIX] 유저 피드 조회 시 전체 피드 개수 반환 쿼리문 수정

+++ src/main/java/org/websoso/WSSServer/repository/FeedCustomRepository.java
@@ -18,7 +18,7 @@ public interface FeedCustomRepository {
 
     Slice<Feed> findRecommendedFeeds(Long lastFeedId, Long userId, PageRequest pageRequest, List<Genre> genres);
 
-    int countVisibleFeeds(User owner, Long lastFeedId, Boolean isVisible,
-                          Boolean isUnVisible, List<Genre> genres,
-                          Long visitorId);
+    Long countVisibleFeeds(User owner, Long lastFeedId, Boolean isVisible,
+                           Boolean isUnVisible, List<Genre> genres,
+                           Long visitorId);
 }
