+++ src/main/java/org/websoso/WSSServer/service/FeedService.java
@@ -488,7 +488,7 @@ public class FeedService {
 
             // TODO Slice의 hasNext()로 판단하도록 수정
             Boolean isLoadable = visibleFeeds.size() == size;
-            int feedsCount = feedRepository.countVisibleFeeds(owner, lastFeedId, isVisible,
+            long feedsCount = feedRepository.countVisibleFeeds(owner, lastFeedId, isVisible,
                     isUnVisible, genres, visitorId);
 
             return UserFeedsGetResponse.of(isLoadable, feedsCount, userFeedGetResponseList);
